fix test: Use old not renovated scrollable instead of new renovated
