+++ js/renovation/components.js
@@ -5,8 +5,8 @@ module.exports = [
     { name: 'Button', pathInRenovationFolder: 'ui/button.j', pathInJSFolder: 'ui/button.js' },
     { name: 'CheckBox', pathInRenovationFolder: 'ui/editors/check_box/check_box.j', pathInJSFolder: 'ui/check_box.js' },
     // { name: 'Widget', pathInRenovationFolder: 'ui/widget.j' },
-    { name: 'ScrollView', pathInRenovationFolder: 'ui/scroll_view/scroll_view.j', pathInJSFolder: 'ui/scroll_view.js', pathToWrapper: '/testing/helpers/renovationScrollViewHelper.js', inProgress: true },
-    { name: 'Scrollable', pathInRenovationFolder: 'ui/scroll_view/scrollable.j', pathInJSFolder: 'ui/scroll_view/ui.scrollable.js', pathToWrapper: '/testing/helpers/renovationScrollableHelper.js', inProgress: true },
+    { name: 'ScrollView', pathInRenovationFolder: 'ui/scroll_view/scroll_view.j', pathInJSFolder: 'ui/scroll_view.js', pathToWrapper: '/testing/helpers/renovationScrollViewHelper.js' },
+    { name: 'Scrollable', pathInRenovationFolder: 'ui/scroll_view/scrollable.j', pathInJSFolder: 'ui/scroll_view/ui.scrollable.js', pathToWrapper: '/testing/helpers/renovationScrollableHelper.js' },
     // { name: 'DataGrid', pathInRenovationFolder: 'ui/grids/data_grid/data_grid.j', pathInJSFolder: 'ui/data_grid.js' },
     { name: 'DataGridNext', pathInRenovationFolder: 'ui/grids/data_grid_next/data_grid_next.j' },
     { name: 'DataGridNextPaging', pathInRenovationFolder: 'ui/grids/data_grid_next/paging/paging.j' },
